requirements.txt
pyinstaller
websocket-server
websocket-client==0.48.0
psutil